removed points with no opportunities and population from the maps
there was a specific point that had neither opportunities nor population
(but had R$ 23 total income) that looked strange in the difference map
in the all modes scenario. I had a good look trying to find the issue,
but couldn't find any. as far as I know it's correct, but the result
is weird. but this commit removes this point (and some other as well)
from the maps and visualisations because it doesn't affect the results
(since all analysis are weighted by population, which is 0)